+++ brainscore_vision/data/geirhos2021/__init__.py
@@ -2,7 +2,7 @@ import pandas as pd
 from brainio.assemblies import BehavioralAssembly
 
 from brainscore_vision import data_registry, stimulus_set_registry
-from brainscore_vision.utils.s3 import load_assembly_from_s3, load_stimulus_set_from_s3
+from brainscore_vision.data_helpers.s3 import load_assembly_from_s3, load_stimulus_set_from_s3
 from brainscore_vision.data_helpers.helper import version_id_df, build_filename
 
 BIBTEX = """@article{geirhos2021partial,
