+++ brainscore_vision/data/seibert2019/__init__.py
@@ -1,7 +1,7 @@
 from brainio.assemblies import NeuronRecordingAssembly
 
 from brainscore_vision import data_registry
-from brainscore_vision.utils.s3 import load_assembly_from_s3
+from brainscore_vision.data_helpers.s3 import load_assembly_from_s3
 from brainscore_vision.data_helpers.helper import version_id_df, build_filename
 
 BIBTEX = """@article{lee2019sensitivity,
